Check if module is a submodule.
When you inspect a module in a package, make sure the package name is in
the submodule name.  It you *don't* do this, then you will follow the
system packages down.  I have no idea what will happen if you pass in a
system module to this.  I know for my dissertation work I had to check
for package members because I wound up recursing through the system
modules.